+++ src/lib/bitmap.c
@@ -14,10 +14,10 @@ ssize_t bitmap_find_nth(bitmap_t* map, size_t size, size_t nth, size_t start, bo
     size_t count = 0;
     unsigned bit = set ? 1 : 0;
 
-    for (ssize_t i = start; i < size; i++) {
+    for (size_t i = start; i < size; i++) {
         if (bitmap_get(map, i) == bit) {
             if (++count == nth) {
-                return i;
+                return (ssize_t)i;
             }
         }
     }
@@ -69,10 +69,10 @@ ssize_t bitmap_find_consec(bitmap_t* map, size_t size, size_t start, size_t n, b
         return -1;
     }
 
-    while (i < size) {
+    while (i < (ssize_t)size) {
         // find the last (with n as maximum) contiguous set page
         count = bitmap_count_consecutive(map, size, i, n);
-        if (count < n) { // if didn't found enough n contiguous set pages
+        if (count < (ssize_t)n) { // if didn't found enough n contiguous set pages
             i += count;
             // find the last contiguous ~set page
             i += bitmap_count_consecutive(map, size, i, -1);
@@ -81,7 +81,7 @@ ssize_t bitmap_find_consec(bitmap_t* map, size_t size, size_t start, size_t n, b
         }
     }
 
-    if (i >= size) {
+    if (i >= (ssize_t)size) {
         i = -1;
     }
